+++ src/uci.rs
@@ -66,7 +66,7 @@ pub fn uci_loop() {
 
         match tokens[0] {
             "uci" => {
-                println!("id name clrsrc 1.0.1");
+                println!("id name clrsrc 1.0.2");
                 println!("id author clrsrc contributors");
                 println!("option name Hash type spin default 64 min 1 max 65536");
                 println!("option name Threads type spin default 1 min 1 max 256");